# No count in bit writer