This PR extends internal `low_value` buffer of bit writer and sends as many bytes as possible when number of stream bits in `low_value` exceeds MAX_STREAM_BITS. It gives substantial performance gain due to much better branch prediction on `if count >= 0` - now it is very rarely true.

Current `main`  7ee30b625d3ef7e3019345bc17d5d510395d1f8a
```
2024-11-16T21:36:50.846Z INFO  [lepton_jpeg::structs::lepton_file_writer] compressing to Lepton format
2024-11-16T21:36:51.329Z INFO  [lepton_jpeg::structs::lepton_file_writer] Number of threads: 8
2024-11-16T21:36:52.735Z INFO  [lepton_jpeg::structs::lepton_file_writer] worker threads 9667ms of CPU time in 1405ms of wall time
2024-11-16T21:36:52.735Z INFO  [lepton_jpeg::structs::lepton_file_writer] decompressing to verify contents
2024-11-16T21:36:54.336Z INFO  [lepton_jpeg_util] compressed input 22171278, output 17324076 bytes (compression = 28.0%)
2024-11-16T21:36:54.336Z INFO  [lepton_jpeg_util] Main thread CPU: 3489ms, Worker thread CPU: 20666 ms, walltime: 3489 ms

 Performance counter stats for 'taskset -c 10 nice -n -20 target/release/lepton_jpeg_util images/img_52MP_7k.jpg images/img_52MP_7k2.lep':

       872 864 510      cache-references                                                        (42,03%)
        81 345 085      cache-misses                     #    9,32% of all cache refs           (42,20%)
    15 679 683 044      cycles                                                                  (41,93%)
       844 619 245      ic_fetch_stall.ic_stall_back_pressure                                        (41,82%)
     1 127 999 657      stalled-cycles-frontend          #    7,19% frontend cycles idle        (41,57%)
    38 437 122 047      instructions                     #    2,45  insn per cycle            
                                                  #    0,03  stalled cycles per insn     (41,59%)
     4 360 833 730      branch-instructions                                                     (41,79%)
       164 653 082      branch-misses                    #    3,78% of all branches             (41,85%)
     5 188 234 309      ic_fetch_stall.ic_stall_any                                             (42,05%)
        36 892 099      ic_fetch_stall.ic_stall_dq_empty                                        (42,17%)
        68 333 545      l2_cache_misses_from_ic_miss                                            (42,13%)
     2 117 822 829      l2_latency.l2_cycles_waiting_on_fills                                        (42,01%)
           184 004      faults                                                                
                 1      migrations                                                            

       3,523626815 seconds time elapsed

       3,201068000 seconds user
       0,320906000 seconds sys
```
This PR 0426f91e1f2f98bcea07f7983fdb85f5cad3a2e5
```
2024-11-16T21:38:12.721Z INFO  [lepton_jpeg::structs::lepton_file_writer] compressing to Lepton format
2024-11-16T21:38:13.198Z INFO  [lepton_jpeg::structs::lepton_file_writer] Number of threads: 8
2024-11-16T21:38:14.474Z INFO  [lepton_jpeg::structs::lepton_file_writer] worker threads 8782ms of CPU time in 1274ms of wall time
2024-11-16T21:38:14.474Z INFO  [lepton_jpeg::structs::lepton_file_writer] decompressing to verify contents
2024-11-16T21:38:16.077Z INFO  [lepton_jpeg_util] compressed input 22171278, output 17324074 bytes (compression = 28.0%)
2024-11-16T21:38:16.077Z INFO  [lepton_jpeg_util] Main thread CPU: 3356ms, Worker thread CPU: 19835 ms, walltime: 3356 ms

 Performance counter stats for 'taskset -c 10 nice -n -20 target/release/lepton_jpeg_util images/img_52MP_7k.jpg images/img_52MP_7k2.lep':

       841 950 286      cache-references                                                        (42,16%)
        74 634 666      cache-misses                     #    8,86% of all cache refs           (41,91%)
    15 098 618 381      cycles                                                                  (41,72%)
       915 407 820      ic_fetch_stall.ic_stall_back_pressure                                        (41,58%)
       986 670 263      stalled-cycles-frontend          #    6,53% frontend cycles idle        (41,57%)
    38 910 234 032      instructions                     #    2,58  insn per cycle            
                                                  #    0,03  stalled cycles per insn     (41,53%)
     4 265 509 081      branch-instructions                                                     (41,82%)
       147 881 884      branch-misses                    #    3,47% of all branches             (42,01%)
     5 396 121 804      ic_fetch_stall.ic_stall_any                                             (42,07%)
        40 630 197      ic_fetch_stall.ic_stall_dq_empty                                        (42,20%)
        61 276 546      l2_cache_misses_from_ic_miss                                            (42,38%)
     2 181 284 237      l2_latency.l2_cycles_waiting_on_fills                                        (42,33%)
           184 036      faults                                                                
                 1      migrations                                                            

       3,391659429 seconds time elapsed

       3,064401000 seconds user
       0,323831000 seconds sys
```