@Melirius 2024-11-16
As in this variant of buffer it is difficult to achieve the same behavior as initial Lepton implementation on very small buffers (specifically before minimal stream length is 4 bytes, while here it can be shorter writing down only real stream bytes) - hashes of streams in two tests are changed.