@Melirius 2024-11-16
Taking into account that decoding of the formed Lepton file takes $7.3\times10^9$ cycles, encoding here takes $7.8\times10^9$ cycles that is only 7 % slower.