@Melirius 2024-11-17
Now it is even faster:
```
2024-11-17T09:12:57.266Z INFO  [lepton_jpeg::structs::lepton_file_writer] compressing to Lepton format
2024-11-17T09:12:57.747Z INFO  [lepton_jpeg::structs::lepton_file_writer] Number of threads: 8
2024-11-17T09:12:58.993Z INFO  [lepton_jpeg::structs::lepton_file_writer] worker threads 8504ms of CPU time in 1246ms of wall time
2024-11-17T09:12:58.993Z INFO  [lepton_jpeg::structs::lepton_file_writer] decompressing to verify contents
2024-11-17T09:13:00.611Z INFO  [lepton_jpeg_util] compressed input 22171278, output 17324074 bytes (compression = 28.0%)
2024-11-17T09:13:00.611Z INFO  [lepton_jpeg_util] Main thread CPU: 3344ms, Worker thread CPU: 19654 ms, walltime: 3344 ms

 Performance counter stats for 'taskset -c 10 nice -n -20 target/release/lepton_jpeg_util images/img_52MP_7k.jpg images/img_52MP_7k2.lep':

       860 171 569      cache-references                                                        (41,57%)
        79 969 138      cache-misses                     #    9,30% of all cache refs           (41,63%)
    15 030 765 275      cycles                                                                  (41,74%)
       565 084 589      ic_fetch_stall.ic_stall_back_pressure                                        (42,03%)
     1 035 372 456      stalled-cycles-frontend          #    6,89% frontend cycles idle        (42,18%)
    37 563 953 194      instructions                     #    2,50  insn per cycle            
                                                  #    0,03  stalled cycles per insn     (42,13%)
     4 192 550 558      branch-instructions                                                     (42,19%)
       148 847 381      branch-misses                    #    3,55% of all branches             (42,17%)
     5 334 304 010      ic_fetch_stall.ic_stall_any                                             (42,11%)
        40 808 153      ic_fetch_stall.ic_stall_dq_empty                                        (41,96%)
        65 332 752      l2_cache_misses_from_ic_miss                                            (41,88%)
     2 049 967 756      l2_latency.l2_cycles_waiting_on_fills                                        (41,67%)
           184 019      faults                                                                
                 1      migrations                                                            

       3,380048890 seconds time elapsed

       3,042535000 seconds user
       0,335838000 seconds sys
```