@mcroomp 2024-11-18
That's great! I'll need to test with the corpus to make sure that all the files are decodablet that will take a couple days.