@Melirius 2024-11-19
Using the same scheme of making checks more rare as in reader, I've got even to
```
2024-11-19T21:00:57.665Z INFO  [lepton_jpeg::structs::lepton_file_writer] compressing to Lepton format
2024-11-19T21:00:58.129Z INFO  [lepton_jpeg::structs::lepton_file_writer] Number of threads: 8
2024-11-19T21:00:59.328Z INFO  [lepton_jpeg::structs::lepton_file_writer] worker threads 8281ms of CPU time in 1198ms of wall time
2024-11-19T21:00:59.328Z INFO  [lepton_jpeg::structs::lepton_file_writer] decompressing to verify contents
2024-11-19T21:01:00.907Z INFO  [lepton_jpeg_util] compressed input 22171278, output 17324074 bytes (compression = 28.0%)
2024-11-19T21:01:00.907Z INFO  [lepton_jpeg_util] Main thread CPU: 3241ms, Worker thread CPU: 19155 ms, walltime: 3241 ms

 Performance counter stats for 'taskset -c 10 nice -n -20 target/release/lepton_jpeg_util images/img_52MP_7k.jpg images/img_52MP_7k2.lep':

       841 064 356      cache-references                                                        (42,25%)
        78 942 226      cache-misses                     #    9,39% of all cache refs           (41,95%)
    14 732 203 268      cycles                                                                  (41,82%)
       826 513 797      ic_fetch_stall.ic_stall_back_pressure                                        (41,74%)
     1 007 162 397      stalled-cycles-frontend          #    6,84% frontend cycles idle        (41,64%)
    36 880 688 419      instructions                     #    2,50  insn per cycle            
                                                  #    0,03  stalled cycles per insn     (41,77%)
     4 299 054 828      branch-instructions                                                     (41,94%)
       158 781 346      branch-misses                    #    3,69% of all branches             (41,92%)
     5 091 345 293      ic_fetch_stall.ic_stall_any                                             (42,02%)
        36 590 376      ic_fetch_stall.ic_stall_dq_empty                                        (42,07%)
        67 544 871      l2_cache_misses_from_ic_miss                                            (42,14%)
     1 892 218 221      l2_latency.l2_cycles_waiting_on_fills                                        (42,28%)
           184 062      faults                                                                
                 1      migrations                                                            

       3,275811568 seconds time elapsed

       2,962062000 seconds user
       0,311901000 seconds sys
```
After merging this PR I prepare another with these improvements.