@Melirius 2024-11-20
@mcroomp, any news about corpus test?