@mcroomp 2024-11-21
Looks good... looks like it produces the same bits as before (except for one byte less sometimes at the end).